Reverted removal of enableProgTh

+++ src/bindings/python/nixl_bindings.cpp
@@ -348,7 +348,7 @@ PYBIND11_MODULE(_bindings, m) {
         .def(py::init<bool, bool>())
         .def(py::init<bool, bool, int>())
         .def(py::init<bool, bool, int, nixl_thread_sync_t>())
-        .def(py::init<bool, bool, int, nixl_thread_sync_t, uint16_t>());
+        .def(py::init<bool, bool, int, nixl_thread_sync_t, size_t>());
 
     //note: pybind will automatically convert notif_map to python types:
     //so, a Dictionary of string: List<string>
